ai-service/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
PyPDF2
spacy
nltk
